nuPlan/requirements_nuplan.txt
aioboto3
aiofiles
bokeh==2.4.3
boto3==1.24.59
cachetools
casadi
control==0.9.1
coverage
docker
Fiona
geopandas>=0.12.1
grpcio==1.43.0
grpcio-tools==1.43.0
guppy3==3.1.2
hydra-core==1.1.0
hypothesis
joblib
jupyter
jupyterlab
matplotlib==3.9.2
mock
moto
nest_asyncio
numpy==1.23.4
opencv-python<=4.5.1.48
pandas
Pillow
pre-commit
psutil
pyarrow
pyinstrument
pyogrio
pyquaternion>=0.9.5
pytest
rasterio
ray
requests
retry
rtree
s3fs
scipy
selenium
setuptools==59.5.0
Shapely>=2.0.0
SQLAlchemy==1.4.27
sympy
testbook
tornado
tqdm
typer
ujson
urllib3
